Refactor: Move LLMSampleSelector class to a separate file
To avoid potential git conflicts, make the code easier to maintain and align with single-responsibility principle.

Also made two public functions private since there are not used anywhere else.
